// src/infrastructure/api/generated/model/petOwnershipResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Pet Rescue API
 * RESTful API for Pet Rescue and Adoption Platform
 * OpenAPI spec version: 1.0.0
 */
import type { PetOwnershipResponseDtoOwnerType } from './petOwnershipResponseDtoOwnerType';

/**
 * Pet ownership history entry
 */
export interface PetOwnershipResponseDto {
  petId?: string;
  ownerType?: PetOwnershipResponseDtoOwnerType;
  ownerId?: string;
  ownerName?: string;
  fromTime?: string;
  toTime?: string;
}