// src/infrastructure/api/generated/model/petSummaryResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Pet Rescue API
 * RESTful API for Pet Rescue and Adoption Platform
 * OpenAPI spec version: 1.0.0
 */
import type { OrganizationMinimalDto } from './organizationMinimalDto';
import type { PetOwnerSummaryDto } from './petOwnerSummaryDto';
import type { PetSummaryResponseDtoGender } from './petSummaryResponseDtoGender';
import type { PetSummaryResponseDtoHealthStatus } from './petSummaryResponseDtoHealthStatus';
import type { PetSummaryResponseDtoStatus } from './petSummaryResponseDtoStatus';

/**
 * Pet summary for list views
 */
export interface PetSummaryResponseDto {
  petId?: string;
  petCode?: string;
  name?: string;
  species?: string;
  breed?: string;
  /** Age in months */
  age?: number;
  ageDisplay?: string;
  vaccinated?: boolean;
  gender?: PetSummaryResponseDtoGender;
  status?: PetSummaryResponseDtoStatus;
  healthStatus?: PetSummaryResponseDtoHealthStatus;
  owner?: PetOwnerSummaryDto;
  organization?: OrganizationMinimalDto;
  province?: string;
  provinceCode?: number;
  ward?: string;
  wardCode?: number;
  /** Thumbnail image URL (first image) */
  imageUrl?: string;
}