// src/infrastructure/api/generated/model/createPetRequestDtoGender.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Pet Rescue API
 * RESTful API for Pet Rescue and Adoption Platform
 * OpenAPI spec version: 1.0.0
 */

export type CreatePetRequestDtoGender = typeof CreatePetRequestDtoGender[keyof typeof CreatePetRequestDtoGender];


export const CreatePetRequestDtoGender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  UNKNOWN: 'UNKNOWN',
} as const;